chore: reorganize unit tests in the redis-v5 package
